chore: moved view sub-header to app header
