+++ web/components/views/index.ts
@@ -5,3 +5,4 @@ export * from "./quick-actions";
 export * from "./view-list-item";
 export * from "./views-list";
 export * from "./view-list-item-action";
+export * from "./view-list-header";
